Fix grecaptcha.render is not a function
Google initializes the grecaptcha object with only the ready function at first. The remaining functions, including render and execute, are added to the object at a later point. Thus, the interval check for just the object is insufficient and introduces a race condition. Adding a check for the render function itself resolves this.